# internal/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.CreateUserRequest:
    properties:
      email:
        type: string
      is_admin:
        type: boolean
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  handlers.GrantPermissionRequest:
    properties:
      role_id:
        type: integer
      user_id:
        type: string
      workspace_id:
        type: string
    required:
    - role_id
    - user_id
    - workspace_id
    type: object
...
  models.AuditLog:
    properties:
      action:
        description: e.g., "create_workspace", "grant_permission"
        type: string
      details_json:
        description: Additional context in JSON
        type: string
      id:
        type: integer
      resource:
        description: e.g., "workspace:123", "user:456"
        type: string
      timestamp:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
  models.FederatedIdentity:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: string
      issuer:
        type: string
      name:
        type: string
      subject:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
  models.FederatedIdentityReview:
    properties:
      avatar_url:
        type: string
      collision_field:
        type: string
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: string
      issuer:
        type: string
      name:
        type: string
      reviewed_at:
        type: string
      reviewed_by:
        type: string
      status:
        type: string
      subject:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
      username:
        type: string
    type: object
  models.Group:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      source:
        $ref: '#/definitions/models.GroupSource'
      updated_at:
        type: string
    type: object
...
  models.GroupSource:
    enum:
    - native
    - oidc
    type: string
    x-enum-varnames:
    - GroupSourceNative
    - GroupSourceOIDC
...
  models.Permission:
    properties:
      created_at:
        type: string
      id:
        type: integer
      role:
        $ref: '#/definitions/models.Role'
      role_id:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
      workspace:
        $ref: '#/definitions/models.Workspace'
      workspace_id:
        type: string
    type: object
  models.Role:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  models.Workspace:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      owner:
        $ref: '#/definitions/models.User'
      owner_id:
        type: string
      package_manager:
        description: '"pixi" or "uv"'
        type: string
      path:
        description: filesystem path (local-mode)
        type: string
      size_bytes:
        type: integer
      source:
        description: '"managed", "local"'
        type: string
      status:
        $ref: '#/definitions/models.WorkspaceStatus'
      updated_at:
        type: string
    type: object
  models.WorkspaceStatus:
    enum:
    - pending
    - creating
    - ready
    - failed
    - deleting
    type: string
    x-enum-varnames:
    - WsStatusPending
    - WsStatusCreating
    - WsStatusReady
    - WsStatusFailed
    - WsStatusDeleting
...
  service.DashboardStats:
    properties:
      total_disk_usage_bytes:
        type: integer
      total_disk_usage_formatted:
        type: string
    type: object
...
  service.UserWithAdmin:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_admin:
        type: boolean
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8460
info:
  contact: {}
  description: Multi-User Workspace Management System API
  title: Nebi API
  version: "1.0"
paths:
  /admin/audit-logs:
    get:
      parameters:
      - description: Filter by user ID
        in: query
        name: user_id
        type: string
      - description: Filter by action
        in: query
        name: action
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AuditLog'
            type: array
      security:
      - BearerAuth: []
      summary: List audit logs
      tags:
      - admin
  /admin/dashboard/stats:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.DashboardStats'
      security:
      - BearerAuth: []
      summary: Get admin dashboard statistics
      tags:
      - admin
  /admin/federated-identity-reviews:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.FederatedIdentityReview'
            type: array
      security:
      - BearerAuth: []
      summary: List federated identity reviews
      tags:
      - admin
  /admin/federated-identity-reviews/{id}/approve:
    post:
      parameters:
      - description: Review UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.FederatedIdentity'
      security:
      - BearerAuth: []
      summary: Approve a pending federated identity review
      tags:
      - admin
  /admin/federated-identity-reviews/{id}/reject:
    post:
      parameters:
      - description: Review UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Reject a pending federated identity review
      tags:
      - admin
...
  /admin/groups/{id}/grant-admin:
    delete:
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Revoke admin from a group (admin only)
      tags:
      - admin
    post:
      description: Every current and future member of the group gains effective admin
        via Casbin role inheritance.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Promote a group to admin (admin only)
      tags:
      - admin
...
  /admin/permissions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Permission'
            type: array
      security:
      - BearerAuth: []
      summary: List all permissions
      tags:
      - admin
    post:
      consumes:
      - application/json
      parameters:
      - description: Permission details
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/handlers.GrantPermissionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Permission'
      security:
      - BearerAuth: []
      summary: Grant workspace access to a user
      tags:
      - admin
  /admin/permissions/{id}:
    delete:
      parameters:
      - description: Permission ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Revoke a permission
      tags:
      - admin
...
  /admin/roles:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Role'
            type: array
      security:
      - BearerAuth: []
      summary: List all roles
      tags:
      - admin
  /admin/users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/service.UserWithAdmin'
            type: array
      security:
      - BearerAuth: []
      summary: List all users (admin only)
      tags:
      - admin
    post:
      consumes:
      - application/json
      parameters:
      - description: User details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
      security:
      - BearerAuth: []
      summary: Create a new user (admin only)
      tags:
      - admin
  /admin/users/{id}:
    delete:
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete a user (admin only)
      tags:
      - admin
    get:
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.UserWithAdmin'
      security:
      - BearerAuth: []
      summary: Get user by ID (admin only)
      tags:
      - admin
  /admin/users/{id}/groups:
    get:
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Group'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List the groups a user belongs to (admin only)
      tags:
      - admin
  /admin/users/{id}/toggle-admin:
    post:
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.UserWithAdmin'
      security:
      - BearerAuth: []
      summary: Toggle admin status for a user
      tags:
      - admin
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"